# RavenDB-22986 - Free space searching improvements using in-memory state