### Issue link

https://issues.hibernatingrhinos.com/issue/RavenDB-22986/Free-space-fragmentation

### Additional description

This is based on the optimization that was done here: https://github.com/ravendb/ravendb/pull/19392/commits/267600dae179ccfb00f3a8c72bab82de82ed71f9#diff-d2dbb8e40687614174e0654344ec0c25a05c45ec41bbe0df0c9b42df8d7c35f2R193

- Remember the maximum number of pages that we **couldn't** allocate from a segment, and skip checking it if it's known **not** to have the required consecutive range.
- On FreePage, we simply delete the information for that segment (this requires a re-check, and we'll remember it again if needed). The re-check was optimized here: https://github.com/ravendb/ravendb/pull/19386.
- On RollBack, we clear all information and require a re-check (this could be turned into a proper state on the environment record, but hasn't been done so far to make it easier to merge into previous versions).
- This approach keeps the state minimal and is essentially self-healing if necessary.

**Results for real life highly fragmented sections**
The tests simulate a single batch of patch operations with a batch size of 1024, targeting sizes that are not available in any of the 5000 sections (Except a size of 1 which **does** exist).

The reason for the different numbers in [this test](https://github.com/ravendb/ravendb/pull/19386) is that it was conducted at the storage level, whereas the previous test was performed directly on the `StreamBitArray`.

Requested size: 1
original: 2ms
current: 1ms
new: 1ms

Requested size: 2-31
original: 20,754ms
current: 2,199ms
new: 109ms

Requested size: 32-63
original: 18,927ms
current: 2,230ms
new: 102ms

Requested size: 32-2048
original: 2,097ms
current: 816ms
new: 102ms

Requested size: 1-2048
original: 2,353ms
current: 849ms
new: 113ms

### Type of change

- [ ] Bug fix
- [ ] Regression bug fix
- [x] Optimization
- [ ] New feature

### How risky is the change?

- [x] Low 
- [ ] Moderate 
- [ ] High
- [ ] Not relevant

### Backward compatibility

- [x] Non breaking change
- [ ] Ensured. Please explain how has it been implemented?
- [ ] Breaking change
- [ ] Not relevant

### Is it platform specific issue?

- [ ] Yes. Please list the affected platforms.
- [x] No

### Documentation update

- [ ] This change requires a documentation update. Please mark the issue on YouTrack using `Documentation Required` tag.
- [x] No documentation update is needed 

### Testing by Contributor

- [x] Tests have been added that prove the fix is effective or that the feature works
- [ ] Internal classes added to the test class (e.g. entity or index definition classes) have the lowest possible access modifier (preferable `private`) 
- [ ] It has been verified by manual testing

### Testing by RavenDB QA team

- [ ] This change requires a special QA testing due to possible performance or resources usage implications (CPU, memory, IO). Please mark the issue on YouTrack using `QA Required` tag.
- [x] No special testing by RavenDB QA team is needed

### Is there any existing behavior change of other features due to this change?

- [ ] Yes. Please list the affected features/subsystems and provide appropriate explanation
- [x] No

### UI work

- [ ] It requires further work in the Studio. Please mark the issue on YouTrack using `Studio Required` tag.
- [x] No UI work is needed