@ppekrol 2024-11-22
@grisha-kotler can you paste your benchmark? Like in the prev PR